% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Title}
\usage{
myncurve(mu, sigma, color, a)
}
\arguments{
\item{mu}{-  blah blah}

\item{sigma}{- blooh blooh}
}
\value{
- uhhh
}
\description{
Title
}
\examples{
- myncurve(1, 2, 4, "blue")
}
